submitoop 2.0
Change the concept behind my submitoop to now create the command for the specific command you wish to create (ie at home, in lab or DIY) You will have to adjust the command to your profs name and section but the program will prompt you for the week each time.
After that copy-paste the command and hit enter and the submitter will initiate.